my_finetune/convert_to_gguf.py
import argparse
from llama_cpp import Llama
from transformers import AutoTokenizer, AutoModelForCausalLM
import torch


def main():
    parser = argparse.ArgumentParser(description="Convert fine-tuned model to GGUF")
    parser.add_argument("--model_path", type=str, required=True, help="Path to fine-tuned model")
    parser.add_argument("--output_file", type=str, required=True, help="Output GGUF file")
    parser.add_argument("--quantization", type=str, default="q4_k_m", help="Quantization method (q4_k_m, q5_k_m, q8_0)")
    
    args = parser.parse_args()
    
    print(f"Converting model from {args.model_path} to {args.output_file}")
    print(f"Quantization: {args.quantization}")
    
    # Note: For proper GGUF conversion, you need to use the llama.cpp convert.py script
    # This is a simplified version that loads the model for LM Studio compatibility
    # For full conversion, use: https://github.com/ggerganov/llama.cpp/blob/master/convert.py
    
    print("\nNote: For proper GGUF conversion, use the llama.cpp convert.py script:")
    print("1. Clone llama.cpp: git clone https://github.com/ggerganov/llama.cpp")
    print("2. Run: python convert.py {args.model_path} --outfile {args.output_file} --outtype {args.quantization}")
    print("\nAlternatively, LM Studio can load Hugging Face models directly.")
    print("Just point LM Studio to: {args.model_path}")
    
    # For now, we'll create a simple placeholder script that shows the path
    with open(args.output_file + ".txt", "w") as f:
        f.write(f"Model path for LM Studio: {args.model_path}\n")
        f.write(f"To convert to GGUF, use llama.cpp convert.py\n")
    
    print(f"\nCreated {args.output_file}.txt with instructions")
    print(f"You can load the model directly in LM Studio from: {args.model_path}")


if __name__ == "__main__":
    main()
