my_finetune/requirements.txt
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
peft>=0.6.0
datasets>=2.14.0
trl>=0.7.0
llama-cpp-python>=0.2.0
sentencepiece>=0.1.99
protobuf>=4.25.0
